springboot-parent/springboot-jersey/src/main/java/com/hsiao/springboot/jersey/ServletInitializer.java
package com.hsiao.springboot.jersey;


import org.springframework.boot.builder.SpringApplicationBuilder;
import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;

/**
 *
 * 〈一句话功能简述〉<br>
 *
 * @projectName springboot-parent
 * @title: ServletInitializer
 * @description: TODO
 * @author xiao
 * @create 2021/8/7
 * @since 1.0.0
 */
public class ServletInitializer extends SpringBootServletInitializer {

    @Override
    public SpringApplicationBuilder configure(SpringApplicationBuilder application) {
        return application.sources(JerseyApplication.class);
    }
}
